Added sim config to expose bug with old KBK channel version.
An old KBK channel mod description (part of the model originally
published on modelDB by Solinas, but then removed in a second release)
has somehow survived and is still present in Golgi_040408_C1, which is
the cell model used in the network simulation. This old version of
the mod file is fine (ie, the cell behaves as intended) only if
simulated with NEURON v<=6.2, but with newer versions of the
simulator, when using fixed-timestep integration, gives rise to an odd
'doublet-spiking' behaviour for the Golgi cell.

In this commit, I have added a new simulation configuration (TestKBK)
that compares two models which are identical except for using the two
different versions of this mod file, so to expose the difference in
spiking pattern that they generate. The TestKBK.py script, for the
moment, just runs this simulation configuration.